#191: Move format functions
